requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

opencv-python==4.7.0.68
Pillow==10.0.0
ultralytics==8.3.48
